requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
PyJWT==2.8.0
pytest==8.3.5
Werkzeug==3.0.6
psycopg2-binary==2.9.6
zipp>=3.19.1
Flask-Limiter==3.5.0
gunicorn==21.2.0